requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.13.0
CurrencyConverter==0.17.9
